Add files via upload
This update makes an improvement to the Cicada language:  a void member can now be used to define other objects.  This means that members that are intended only as data types or class definitions, not objects, can be defined void to avoid the awkwardness of their having an instantiated object.  Specifically, if there's no source variable to provide a type specification when defining an object, the source member's type spec will now be used instead.  It's technically possible for this change to cause an incompatibility with previous code.. but unlikely.


Bugs fixed:

* string tokens could occasionally jam when they shouldn't
* crashes passing void members to certain built-in functions
* errors using >> and : operators weren't reliably thrown and could cause a crash
* member-define didn't work properly when used in multi-define expressions
* problems defining void arrays